\name{BinaryMeth-package}
\alias{BinaryMeth-package}
\docType{package}
\title{
\packageTitle{BinaryMeth}
}
\description{
\packageDescription{BinaryMeth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BinaryMeth}
\packageIndices{BinaryMeth}
}
\author{
\packageAuthor{BinaryMeth}

Maintainer: \packageMaintainer{BinaryMeth}
}
\references{
NA
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }

